Add files via upload
To run this Node.js + Express.js, first we will have to go to the terminal after opening this file in maybe say Visual Studio Code, then to open terminal for this file, first click on index.js, then at the top go to view and click on Terminal. It will open the terminal directly for this file location. 

There in terminal first use "npm init -y" and press enter which will create a node module file, then use "npm i express" so the we can use express features for this file, then use "node index.js" to run the code on localhost:3000 which you will have to type on the browser to see if it if working or not. 

But i used "nodemon index.js" as it will update the localhost whenever we will make any change to the file so we won't have to stop and start it again and again. To use nodemon first we have to use "npm i -g nodemon" in the terminal to install it globally so we can use it next time we want to run a node.js file.